Tracking down these bugs was A PAIN. Better thank me tomorrow. TODO: functions start() and end() in the builder

+++ src/com/exorath/game/api/countdown/CountdownFrame.java
@@ -7,9 +7,9 @@ import com.exorath.game.api.player.GamePlayer;
  */
 public interface CountdownFrame {
 
-    int getDelay();
+    long getDelay();
 
-    int getDuration();
+    long getDuration();
 
     void start(GamePlayer player);
